+++ src/pcelery/runtests.py
@@ -9,13 +9,17 @@ def runtests():
     import sys
     import pytest
     from os import environ
-    from os.path import dirname, join
-    cfg_path = join(dirname(__file__), '..', 'setup.cfg')
+    from pathlib import Path
+    root_dir_path = Path(__file__).parent / '..' / '..'
+    cfg_path = root_dir_path / 'src' / 'setup.cfg'
 
     args = sys.argv[1:]
     if not args or args[0].startswith('-'):
         args += ['--pyargs', 'pcelery']
-    args = ['-c', cfg_path] + args
-    environ['IS_TESTING'] = 'True'
+    args = [
+               '-c', str(cfg_path),
+               '--rootdir', str(root_dir_path),
+           ] + args
+    environ['IS_TESTING'] = 'true'
 
     return pytest.main(args)