+++ src/pcelery/tests/test_task.py
@@ -62,4 +62,4 @@ def test_cli_command():
     sys.argv.extend(old_argv)
 
     assert exc_info.value.args[0] == 0
-    assert std_out.getvalue().strip() == '5.1.2 (sun-harmonics)'
+    assert std_out.getvalue().strip() == '5.3.4 (emerald-rush)'